Refactor: 핸들러 호출 방식으로 변경
